Move the test to the proper package and add a helper to access package private stuff
